% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{add_tof}
\alias{add_tof}
\title{Agregar objetos}
\usage{
add_tof(..., file, env = parent.frame())
}
\arguments{
\item{...}{nombres de los objetos o un vector tipo character con
los nombres de los objetos que serán agregados}

\item{file}{character: nombre del archivo. Si no existe, es creado}

\item{env}{environment donde "viven" los objetos que serán
almacenados; por omisión, el parent.frame}
}
\value{
character: los nombres de los objetos agregados o
\code{NULL} si algún error
}
\description{
Agrega objetos a un archivo sin eliminar los
previamente almacenados en él; pero objetos con el mismo nombre
son remplazados.
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
vi <- 1:3
fi <- tempfile()
save(df, vi, file = fi)

rd <- add_tof(d2, file = fi)
list_off(fi, meta = FALSE)

unlink(fi)
rm(rd, fi, df, d2, vi)
}
