% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fut.r
\name{dots_arg}
\alias{dots_arg}
\title{dot argument}
\usage{
dots_arg(...)
}
\arguments{
\item{...}{arguments}
}
\value{
\code{NULL}, character or integer vector
}
\description{
arguments in ... returned as a character or integer
vector
}
\examples{
\dontrun{
dots_arg(a, b)
dots_arg("a", "b")
dots_arg(c("a", "b"))
dots_arg(1:3)
}
}
\author{
eddy castellón
}
