% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{get_dff}
\alias{get_dff}
\title{Lee - asigna un data.frame}
\usage{
get_dff(x, file)
}
\arguments{
\item{x}{nombre del data frame o character con el nombre}

\item{file}{character: nombre del archivo}
}
\value{
a data.frame o NULL
}
\description{
Lee un data frame y lo asocia con una variable
}
\seealso{
\code{get_dffs} \code{get_off}
}
\author{
eddy castellón
}
