% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{get_off_c}
\alias{get_off_c}
\title{Leer-objeto}
\usage{
get_off_c(x = character(), file = character())
}
\arguments{
\item{x}{character: nombre del objeto.}

\item{file}{character: nombre del archivo}
}
\value{
objeto o \code{NULL}
}
\description{
Lee un objeto almacenado en archivo
}
\details{
Versión simplificada de \code{get_off} con llamada
estándar
}
\examples{
\dontrun{x <- get_off_c("obj", file = "some-file.rda")}
}
\seealso{
\code{get_off}
}
