% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fut.r
\name{is_path}
\alias{is_path}
\title{path maybe}
\usage{
is_path(x)
}
\arguments{
\item{x}{character}
}
\value{
TRUE if x has at least one slash followed by and ended by
an alphanumeric character
}
\description{
test string begins with one o more alphanumeric
characters followed by file path separators, or it begins with
"./" or ".\\" followed by one or more alpanumeric characters;
in both cases ending in alphanumeric character.
}
\examples{
\dontrun{
is_path(".aa/bb")
is_path("aa/bb")
is_path("aa/bb.")
}
}
\keyword{internal}
