% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{list_dff}
\alias{list_dff}
\title{data.frames almacenados}
\usage{
list_dff(file, meta = TRUE)
}
\arguments{
\item{file}{character: nombre del archivo}

\item{meta}{logical: devolver metadatos? (TRUE por defecto)}
}
\value{
character or data.frame
}
\description{
Data frames almacenados en un archivo.
}
\details{
Alias de la función \code{list_off} con argumento
"data.frame" en el parámetro "class".
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
vi <- 1:3
fi <- tempfile()
save(df, d2, vi, file = fi)

list_dff(fi)
unlink(fi)
rm(df, d2, vi, fi)
}
\seealso{
\code{list_off}, \code{meta}
}
\author{
Eddy Castellón
}
