% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{list_off}
\alias{list_off}
\title{Objetos almacenados}
\usage{
list_off(file, meta = TRUE, class = ".")
}
\arguments{
\item{file}{character. Nombre del archivo}

\item{meta}{logical. Si TRUE (default) devuelve valor del atributo
"meta" de cada objeto.}

\item{class}{character. Hace el reporte sólo para objetos de clase
"class". Por omisión, objetos de cualquier clase (class = ".").}
}
\value{
character si meta = FALSE, o data.frame con una columna para
el nombre y otra para los metadatos; \code{NULL} si ocurre un
error.
}
\description{
Nombres de los objetos almacenados en un archivo.
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
vi <- 1:3
fi <- tempfile()
save(df, d2, vi, file = fi)

list_off(fi, meta = FALSE)
list_off(fi, class = "data.frame")

unlink(fi)
rm(df, d2, vi, fi)
}
\author{
Eddy Castellón
}
