% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{meta<-}
\alias{meta<-}
\title{Metadatos - Asignar}
\usage{
meta(x) <- value
}
\arguments{
\item{x}{objeto}

\item{value}{character: metadatos}
}
\description{
Le asigna a un objeto, una breve descripción o
metadatos.
}
\details{
Al objeto se le agrega el atributo "meta" para registrar
una breve descripción (metadatos), o el nombre de un archivo en
el que se hace una descripción más extensa. Ese nombre debe
contener al menos una pleca ("/") para que la función no
interprete el nombre como metadatos. El contenido del archivo
debe apegarse al formato "DCF", explicado en la ayuda de la
función \code{read.dcf}.\preformatted{Es una alternativa de la función \code{comment}.
}
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
meta(df) <- "./metadata.txt"
meta(df) <- "other_path/metadata.txt"

rm(df)
}
\seealso{
\code{comment}, \code{read.dcf}
}
\author{
Eddy Castellón
}
