% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{meta}
\alias{meta}
\title{Metadatos - Leer}
\usage{
meta(x, read_me = FALSE)
}
\arguments{
\item{x}{nombre del objeto}

\item{read_me}{logical. Si TRUE y el atributo "meta" es nombre de
archivo, devuelve el contenido del archivo. FALSE por omisión.}
}
\value{
string o \code{NA} si el objeto no tiene atributo "meta"
}
\description{
Lee el contenido del atributo "meta" o del archivo
cuyo nombre está registrado en el atributo.
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
meta(df)

rm(df)
}
\seealso{
\code{comment}, \code{read.dcf}
}
\author{
Eddy Castellón
}
