% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fut.r
\name{ok_fname}
\alias{ok_fname}
\title{file name}
\usage{
ok_fname(x = character())
}
\arguments{
\item{x}{character; the file's name}
}
\value{
logical
}
\description{
check file's name is valid using the function
\code{file.create} of base R. If any directory in the path chain
doesn't exists, the file's name is invalid.
}
\author{
eddy castellón
}
