% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{read_dff}
\alias{read_dff}
\title{Leer data.frame}
\usage{
read_dff(..., file, env = parent.frame())
}
\arguments{
\item{...}{nombres de los data frames. Si ausente, se leen todos
los que están en el archivo}

\item{file}{character: nombre del archivo}

\item{env}{environment. Por omisión, el \code{parent.frame} (desde
donde se llama la función}
}
\value{
character con los nombres de los data.frame o \code{NULL}
}
\description{
Lee uno o más data.frame almacenados en archivo.
}
\details{
Lee los data.frame y los copia en el \code{environment}
indicado en el parámetro \code{env}. Llama la función
\code{read_off} con argumento "data.frame" en el parámetro
\code{class}.
}
\seealso{
\code{read_off}
}
\author{
eddy castellón
}
