% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{read_off}
\alias{read_off}
\title{Copiar objetos}
\usage{
read_off(..., file = character(), class = ".", env = parent.frame())
}
\arguments{
\item{...}{nombres de los objetos. Si ausente, se copian todos los
objetos.}

\item{file}{character: nombre del archivo}

\item{class}{character: clase de los objetos a copiar. Por omisión,
cualquier clase (\code{class="."}).}

\item{env}{environment: \code{environment} donde serán copiados los
objetos. Por omisión, el ambiente desde donde se llama la
función (\code{parent.frame}}
}
\value{
character: nombre de los objetos copiados o \code{NULL}
}
\description{
Copia los objetos almacenados en un archivo, a un
\code{environment} específico.
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
vi <- 1:3
fi <- tempfile()
save(df, d2, vi, file = fi)

rd <- read_off(file = fi, class = "data.frame")
exists(rd[1])

rd <- read_off(c("d2", "vi"), file = fi)
class(rd[2])

nwe <- new.env()
rd <- read_off(d2, file = fi, env = nwe)
exists("d2")
exists("d2", envir = nwe)

unlink(fi)
rm(df, d2, vi, fi, nwe, rd)
}
\author{
eddy castellón
}
