% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{rm_off}
\alias{rm_off}
\title{Remover - objeto}
\usage{
rm_off(x, file = character(), file2)
}
\arguments{
\item{x}{nombre del objeto}

\item{file}{character: nombre del archivo que almacena objetos.}

\item{file2}{character: nombre del archivo donde se almacenarán los
objetos que no son removidos. Si ausente, los objetos seguirán
almacenados en el mismo archivo.}
}
\value{
character: nombres de los objetos no removidos o
\code{NULL} si algún error durante el proceso
}
\description{
Remueve objetos de un archivo y guarda los restantes
en el mismo o en otro archivo
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
vi <- 1:3
fi <- tempfile()
save(df, d2, vi, file = fi)

rd <- rm_off(vi, file = fi)
rd

unlink(fi)
rm(rd, fi, df, d2, vi)
}
\author{
Eddy Castellón
}
