% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{save_add}
\alias{save_add}
\title{Agregar objetos}
\usage{
save_add(..., file, env = parent.frame())
}
\arguments{
\item{...}{nombres de los objetos o un vector tipo character con
los nombres de los objetos que serán agregados}

\item{file}{character: nombre del archivo. Si no existe, es creado}

\item{env}{environment donde "viven" los objetos que serán
almacenados; por omisión, el parent.frame}
}
\value{
character: los nombres de los objetos agregados o
\code{NULL} si algún error
}
\description{
Alias de \code{add_tof}.
}
\seealso{
\code{add_tof}
}
