% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{save_df.data.frame}
\alias{save_df.data.frame}
\title{Almacenar data.frame}
\usage{
\method{save_df}{data.frame}(x, name = character(), file, metadata = character())
}
\arguments{
\item{x}{nombre del data.frame}

\item{name}{character: nombre con el que será almacenado el
data.frame. Si ausente, se guarda con el mismo nombre.}

\item{file}{character: nombre del archivo.}

\item{metadata}{character: breve descripción o metadatos del
data.frame, si es que este no tiene una. Vea la función
\code{meta}}
}
\value{
character: nombre con el que es almacenado el data.frame, o
\code{NULL} si falla la operación
}
\description{
Agrega un data.frame a un archivo, con el mismo u otro
nombre. Si en el archivo hay almacenado un objeto con nombre
igual, ese objeto será remplazado por el data.frame.
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
vi <- 1:3
fi <- tempfile()
save(df, vi, file = fi)

d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
save_df(d2,
    name = "first3", file = fi,
    metadata = "(lower)uppercase first letters of the alphabet"
)
list_dff(file = fi)

unlink(fi)
rm(df, vi, d2, fi)
}
\seealso{
\code{meta}, \code{add_tof}, \code{save_add}
}
\author{
eddy castellón
}
