% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fut.r
\name{save_ok}
\alias{save_ok}
\title{save}
\usage{
save_ok(...)
}
\arguments{
\item{...}{arguments passed to function save}
}
\value{
logical; FALSE if save with errors
}
\description{
save without errors?
}
\author{
eddy castellón
}
