% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fut.r
\name{try_load}
\alias{try_load}
\title{load file}
\usage{
try_load(x, env = parent.frame())
}
\arguments{
\item{x}{character; file's name}

\item{env}{object environment where objects are loaded;
\code{parent.frame} by default}
}
\value{
character; objects' names or NULL
}
\description{
load a file catching errors
}
\examples{
\dontrun{
try_load("xx.rda")
try_load("xx.rda", env = new.env())
}
}
\author{
eddy castellón
}
