% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{get_dff}
\alias{get_dff}
\title{Lee - asigna un data.frame}
\usage{
get_dff(x, file)
}
\arguments{
\item{x}{nombre del data frame o character con el nombre}

\item{file}{character: nombre del archivo}
}
\value{
data.frame o NULL
}
\description{
Lee data frame almacenado en archivo
}
\examples{
x <- get_dff(aja, "arar.rda")
x <- get_dff("aja", "arar.rda")
}
\seealso{
\code{get_dffs} \code{get_off}
}
\author{
eddy castellón
}
