% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{get_dffs}
\alias{get_dffs}
\title{Lee - asigna data.frames}
\usage{
get_dffs(file, ...)
}
\arguments{
\item{file}{character: nombre del archivo}

\item{...}{nombres de los data.frame o vector tipo character con
los nombres. Si ausente, se leen todos los data.frame en el
archivo.}
}
\value{
data.frame, lista de data frames o \code{NULL}
}
\description{
Lee uno o más data.frame y los asigna a una variable.
}
\details{
Es alias de la función \code{get_off} con argumento
"data.frame" en el parámetro "class", pero con el parámetro
"file" al inicio, para que no sea necesario escribirlo.
}
\examples{
\dontrun{
  list_dff("path/file")
  # "df1", "df2"
  x <- get_dffs("path/file", df2)
  y <- get_dffs("path/file", df2, df1)
}
}
\seealso{
\code{get_off}
}
\author{
eddy castellón
}
