% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobj.r
\name{get_off}
\alias{get_off}
\title{Lee y asigna objetos}
\usage{
get_off(..., file = character(), class = ".")
}
\arguments{
\item{...}{nombres de objetos. Si ausente, lee todos los
almacenados}

\item{file}{character: nombre del archivo}

\item{class}{character: clase de los objetos. Por omisión, de
cualquier clase (\code{class = "."})}
}
\value{
objeto, lista de objectos o \code{NULL}
}
\description{
Lee uno o más objetos almacenados en archivo y los
asigna (bind) a una variable. La variable será un objeto de
clase lista si son leídos dos o más objetos.
}
\examples{
df <- data.frame(x = 1:3, y = 3:1)
meta(df) <- "some metadata"
d2 <- data.frame(a = letters[1:3], b = LETTERS[1:3])
vi <- 1:3
fi <- tempfile()
save(df, d2, vi, file = fi)

rd <- get_off(file = fi)
sapply(rd, class)

rd <- get_off(file = fi, class = "data.frame")
length(rd)
names(rd[[1]])

rd <- get_off(d2, file = fi)
names(rd)

rd <- get_off(vi, file = fi, class = "integer")
length(rd)

unlink(fi)
rm(df, d2, vi, rd, fi)
}
\seealso{
\code{read_off}
}
\author{
Eddy Castellón
}
